added MPI Programmm

+++ code/Metropolis/adaptive_metropolis.c++
@@ -38,7 +38,7 @@ bool adaptive_metropolis(   Lattice &lattice,
                             flt const& T /*temperature*/,
                             flt const& J /*interaction Strength*/,
                             flt const& maxTimeSeconds,
-                            uint const& maxSteps,
+                            u64 const& maxSteps,
                             Spin const& h,
                             Spin const& k,
                             flt const& maxFactor){
