+++ code/Metropolis/metropolis.c++
@@ -27,12 +27,12 @@ bool metropolis(Lattice &lattice,
                 flt const& T /*temperature*/,
                 flt const& J /*interaction Strength*/,
                 flt const& maxTimeSeconds,
-                uint const& maxSteps,
+                u64 const& maxSteps,
                 Spin const& h,
                 Spin const& k,
                 MoveType const& moveType) {
     // Start time and set step counter to 0
-    uint step_count = 0;
+    u64 step_count = 0;
     measure::Timer watch;
     flt beta = Beta(T);
